saga_orchestrator/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.4
kafka-python==2.0.2
requests==2.31.0
django-cors-headers==4.3.1

